Desktop/side project/API/restapi/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
